requirements.txt
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
adk>=0.0.5
